backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
pydantic[email]==2.10.3
pydantic-settings==2.6.1
fpdf2>=2.7.9
